requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
openai==1.6.1
requests==2.31.0
python-dateutil==2.8.2
